refactor: update chatbot types to aichat type file
