+++ src/types/interface.ts
@@ -22,29 +22,3 @@ export interface Character {
   url: string;
   imagePath: string;
 }
-
-export interface ChatbotResponseSection {
-  title: string;
-  content: string;
-  type?: 'info' | 'warning' | 'success' | 'error';
-}
-
-export interface ChatbotResponseTable {
-  headers: string[];
-  rows: string[][];
-  caption?: string;
-}
-
-export interface ChatbotResponseMetadata {
-  hasStructuredContent: boolean;
-  responseLength: number;
-  processingTimestamp: string;
-}
-
-export interface ChatbotResponse {
-  shortAnswer: string;
-  reasoning?: string | null;
-  sections: ChatbotResponseSection[];
-  tables: ChatbotResponseTable[];
-  metadata: ChatbotResponseMetadata;
-}
